moved filtersidebar to the correct file location + added filtercard
